# openapi.yaml
openapi: 3.0.3
info:
  title: Nova Market Auth Service
  description: Internal authentication service contract. Public clients use NovaApiGateway.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/auth/csrf:
    get:
      tags: [Auth]
      operationId: getCsrf
      summary: Get a CSRF token for a verification session
      security: [{ verificationCookie: [] }]
      responses:
        '200':
          description: CSRF token
          content:
            application/json:
              schema:
                type: object
                required: [csrfToken]
                properties: { csrfToken: { type: string } }
        '401': { description: Verification session required }
...
  /api/auth/email/send-verification:
    post:
      tags: [Auth]
      operationId: sendEmailVerification
      summary: Send an email verification code through the email flow
      security: [{ verificationCookie: [], csrfHeader: [] }]
      responses:
        '204': { description: Verification code queued }
        '401': { description: Verification session required }
        '403': { description: Invalid CSRF token }
...
  /api/auth/phone/send-verification:
    post:
      tags: [Auth]
      operationId: sendPhoneVerification
      summary: Ask the mock mobile-money provider to send a phone verification code
      security: [{ verificationCookie: [], csrfHeader: [] }]
      responses:
        '204': { description: Verification request accepted; the OTP is never returned by auth-service }
        '401': { description: Verification session required }
        '403': { description: Invalid CSRF token }
...
  /api/auth/sudo/{channel}:
    post:
      tags: [Auth]
      operationId: sendSudoVerification
      summary: Send a sudo verification code
      security: [{ accessCookie: [], csrfHeader: [] }]
      parameters:
        - name: channel
          in: path
          required: true
          schema: { type: string, enum: [email, phone] }
      responses:
        '204': { description: Sudo verification code queued }
        '400': { description: Unsupported verification channel }
        '403': { description: Invalid CSRF token }
...
  /internal/validate:
    post:
      tags: [Internal]
      operationId: validateAccessToken
      summary: Validate an access token for another service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [access_token]
              properties: { access_token: { type: string } }
      responses:
        '200': { description: Token validation result }
        '401': { description: Invalid token }
...
components:
  securitySchemes:
    verificationCookie:
      type: apiKey
      in: cookie
      name: verification_token
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
...